Refactor game creation to use GameResponseDTO
The Game Service and Game Resolver have been updated to use the newly added GameResponseDTO for the create game method. The relevant parts of the code have been adjusted to accommodate this change, which enhances the flexibility for game response information. This DTO includes id and name properties of the game to be returned upon creation.