Refactor Game model to adjust current player concept
The Game model and related services have been modified to change the concept of the current player. Instead of storing player ID as 'currentPlayerId', it now stores a player's role in the 'currentRole' field. In addition, the 'Player' type and service have been created to support game creation and player assignment process efficiently. This modification offers a more intuitive way of designating the game's current player status and facilitates player management.