Drop 'currentPlayerId' column and add 'currentRole' column in Game model
The 'currentPlayerId' column is dropped from the 'Game' table and replaced with a 'currentRole' column of TEXT type. This changes the way the current player concept is handled in the 'Game' model, as the player's role as current game player is now stored, instead of their ID. This change provides a more intuitive way of managing the game's current player status.